+++ app/routes/components/Popover/changelog.codegen.ts
@@ -3,4 +3,6 @@ export const changelog = ''
   + '- 2025-09-11: Support the prop `onClickOutside` of `Popover.Portal`.\n'
   + '  Refactor it to make it simpler with zustand. Use `autoUpdate` from\n'
   + '  `@floating-ui/react` to handle the position update automatically. etc.\n'
+  + '- 2025-09-22: Add support for disabling the trigger click handler when clicking\n'
+  + '  outside the portal but clicking the trigger as well.\n'
   ;