+++ src/components/Popover/index.ts
@@ -1,3 +1,3 @@
-// Popover from copy-ui @ 2025-09-11
+// Popover from copy-ui @ 2025-09-22
 
 export { Popover } from "./Popover";