+++ src/components/Popover/CHANGELOG.md
@@ -2,3 +2,5 @@
 - 2025-09-11: Support the prop `onClickOutside` of `Popover.Portal`.
   Refactor it to make it simpler with zustand. Use `autoUpdate` from
   `@floating-ui/react` to handle the position update automatically. etc.
+- 2025-09-22: Add support for disabling the trigger click handler when clicking
+  outside the portal but clicking the trigger as well.